+++ arki/runtime.cc
@@ -38,15 +38,7 @@ Config::Config()
     dir_qmacro.init_config_and_env("qmacro", "ARKI_QMACRO");
     if (const char* envdir = getenv("ARKI_SCAN"))
         dir_scan.push_back(envdir);
-    if (const char* envdir = getenv("ARKI_SCAN_GRIB1"))
-        dir_scan.push_back(envdir);
-    if (const char* envdir = getenv("ARKI_SCAN_GRIB2"))
-        dir_scan.push_back(envdir);
     dir_scan.push_back(conf_dir / "scan");
-    dir_scan.push_back(conf_dir / "scan-grib1");
-    dir_scan.push_back(conf_dir / "scan-grib2");
-    dir_scan_bufr.init_config_and_env("scan-bufr", "ARKI_SCAN_BUFR");
-    dir_scan_odimh5.init_config_and_env("scan-odimh5", "ARKI_SCAN_ODIMH5");
 
     if (const char* envfile = getenv("ARKI_IOTRACE"))
         file_iotrace_output = envfile;